dove deep into the layout
i dont know why but when i try to change the margin-top for the nav the  main tag also moves with it. but when i only move the main tag the nav does not move

another one i noticed is when i added a footer and i did the first margin, it doesnt work but the second one does??? what the f. even when i remove it the second one still works

im convinced it that it my computer  lol