conditionals and loops
in conditionals.html:

created a new function just to make sure that I know what I am doing.
i created a var that can help me find an index  (@) for the var louisaStr, then i console.log it.
then i created a condition if the (index != -1) which is saying if they do find that the index I am looking and does NOT equal to -1 (-1 is stating that they did NOT find the @ i was looking for) then we console.log stating the result and and we say that the condition is true

and just bc i wanted to type this but we also create a document method to get the element id (in this case the id for the button that I  also created) and when we onclick the button, we run the function that I created

for loops.html:
this while loop will do it over and over again until the argument has been met