loops are fun, still learning
what I have learned in while loops,

they do things over and over again until the argument has been met

in this case i created a function called doIt and in it i have a var = louisaLoop = 10 and gave it a while loop. so in this loop it is stating that  while  louisaLoop is greater than 0 it will run the function i provided which to console.log a string and to make louisaLoop - 1.

then when the argument does not meet anymore it will console.log another string stating that "Here is the end of the loop"