doItAgain() with if loops
in this section i learned that you can put in an if statement inside the while loop lol.

i think what it does is the while loop is continuous, putting an is statement can make it even more specific for when something happens .

so in this case I created a function that has a var = 10 name aaron. I gave it a while loop if aaron is greater 0 then show the function of an if statement. in this if statement we are saying that if aaron == to 3 to show an alert stating a string and to show aaron and also there will be a decrement of 1.

NOTE: i realized that the placement of the decrement matters b/c i placed it on top of the alert and it actually showed me 2 in instead of 3, crazy.  so  now i know that it is important to have the decrement or increment with the loops

then all else, so once the while loops passes three, i made it console.log a string stating that now current place of aaron along stating that it has a decrement of 1

i hope that i am able to understand loops cause i remember having a hard time understanding this section