put arrays inside arrays
created new function showing that you can put an array inside an array.

one var called dogs is an array of dogs and the other array called cats is an array of cats and the dogs array. then we created an alert displaying the cats array and we also grabbed element by id and changed the innerHTML displaying dogs var