PART 2
okay, so to make the code cleaner, instead of making two different functions, b/c it does the same thing, we just put in an argument.

when you create an argument in your function, you are sending info to the function. kinda acts like a variable b/c it holds info

then when we want to activate the function we get the element by ID and when we click on either left or right button we will run the aFunction() <--- inside the brackets will be our argument.

so in the left or right button we will put aFunction(this.id)

"this" means we are targeting the Id of the rightButt or the leftButt. and b/c this is an object (package of code) we need to set the property to id b/c if we dont we will get an error of null. we will get an error of null b/c in the aFunction we need to (i guess) be more specific on what "this" means by setting the property to .id

kindaa confusing but hopefully i understand it