passing var multiple times
in this part we created a new function that can have multi parameters and pass it on in multiple variables and create an array with it.

in this example we made it easy  to understand by using our name, last, age

created function and used those 3 as our parameters, then  we created a var with name and last. then another var with the age. then the last var is an array that will show the first two arrays. then returning the last var

we created a button in html and created a document object stating when we click on the button it will activate the function we created.  then in the function we will put our ARGUMENT with whatever name, last, age and it will show in the console.log