omg i thought I messed up
NOTE TO SELF: SAVE EVERYTHING, ESPECIALLY IF I MAKE A MISTAKE!

so i created a function called validateForm(), in this function we created two var, first one is the ID firstName and the sec one is the firstName value

then we created an if statement stating that if the value does NOT equal to "Louisa" meaning that if the user types in the textbox and they did NOT put Louisa, the we are going to run a couple of things,...

we are going to console.log saying it is not working, then we are going to grab the target ID (in this case is the <p> tag) and we are going to input a new text (innterHTML)  saying to input Louisa...

then we  are going to change the text color to blue

then we are going to use the select() method on firstName so when the user inputs the wrong name, it will highlight the text so it makes it easier on the user the type in the textbox

so i made the mistake of doing target.select() and what was crazy was that it showed no errors yet it broke my function. so what had happened was that when I input the wrong name it still took me to the website! so yea i just have to remember those little errors..