made a mistake and almost couldnt figure out why..
lol make a couple of mistake here and i almost couldnt figure out why.

first of all i make the mistake of instead of setting everything to context. i did canvas. yikes

then i also made the mistake of setting var context = document.getContext("2nd"), well two things i did wrong here.
this was suppose to be a method for canvas not the document... AND  I also made the mistake of setting the context to "2nd" it was suppose to be "2d" lol
it sucks b/c in the console it was highlighting the .lineWidth() and not the "2nd"

its fine b/c thist just means that I am learning from my mistakes. i make a lot of mistakes