created the onclick() for the buttons
when you click on the start button, in the function i created an if statement.

but before i created that i created two global variables, so outside of the function.
one is notRunning = true
var animation

for the if statement, if notRunning is true then run (set animation equals to) an interval. in the interval we want to set the discoSquare continue every 10th of a sec, then make notRunning = false

all else (or if you click on it again) it will stet an alert

also created an function when you click on the stop
it will stop the interval and make notRunning true again so when you click on start it will run again